Buka: reconstruction-pattern audit and verified source improvements (#59)
* tools: reject reconstructed retail controls before build and delink

* tools: retain exact reviewed parse exceptions during symbol recovery

* reconstruct: use text and opaque-buffer types at resource read boundaries

* reconstruct: recover the complete Miles sample-state owner

* reconstruct: separate pathfinding scratch storage without padding objects

* reconstruct: index real palette levels in bitmap and icon dimming

* reconstruct: size world-map offset arrays to the three scale choices

* audit: record reconstruction-pattern findings for 467 definitions

* reconstruct: use scalar aggregate directory count

* audit: review bitmap compression and map helper patterns

* reconstruct: use const-correct recruitment resource names

* audit: review army, recruitment and exchange function families

* reconstruct: recover scalar serial receive storage

* reconstruct: use scalar remote packet checksums

* audit: prove local switch targets and retain retail vararg exits

* audit: review serial and remote transport functions and bounds

* reconstruct: use native DirectPlay player ID outputs

* reconstruct: recover indexed configuration walk-speed array

* audit: review DirectPlay and cursor functions and speed consumers

* reconstruct: pair visibility-map allocation with array deletion

* audit: review adventure dispatch and prove DirectPlay drain extent

* reconstruct: replace bottom-view unions with widget-pointer arrays

* audit: finish adventure-manager review and prove retail bounds hazards

* reconstruct: recover one unsigned hero patrol owner

* refactor: remove redundant hero level sample const cast

* audit: prove hero formatter capacities from Buka strings

* audit: retain retail hero skill-search trampoline

* audit: review KB functions and apply typed resource and dialog cleanups

* audit: recover cheat flag owner and remove scalar casts

* audit: review game setup and serialization patterns

* audit: finish GAME function review and apply typed text and scratch cleanups

* audit: review town functions and simplify pointer and scratch owners

* audit: recover adjacent town spell-count owner

* audit: review PHILAI turn setup and movement contracts

* audit: review AI purchase selection and remove numeric casts

* audit: review combat values and simplify numeric conversions

* audit: finish PHILAI function review and simplify mana casts

* audit: review army functions and recover unsigned animation owners

* audit: review combat spells and recover effect array owners

* audit: review spell AI and recover typed spell lookup

* review: recover combat-manager cast and table owners

* review: recover combat AI casts and array owners

* audit: review all combat pathfinding functions and simplify speed restore

* audit: review flight movement casts storage and adjacency contracts

* audit: recover COMMAND packet and adjacency owners

* audit: review Castle UI conversions and bounds

* audit: review Campaign casts and retail state bounds

* audit: review HISCORE record and formatter bounds

* audit: review Overview patterns and remove redundant conversions

* audit: review file requester storage and bounds patterns

* review: audit Windows entry messages and menu ownership

* review: audit combat drawing storage control flow and bounds

* audit: recover New Game packet owners and review all functions

* audit: complete setup dialog pattern and bounds review

* audit: recover expansion campaign name owners and review all functions

* reconstruct: audit EVENTS and recover typed flags and hero tails

* audit: finish global data and annotation review coverage

* reconstruct cursor byte owners and retain required dialect casts

* reconstruct executive text as separate character-array owners

* simplify movie and graphics boundaries with complete pattern trials

* reconcile completed reader-interface findings in the audit ledger

* reconstruct: recover signed tagged network payload views

* reconstruct: replace duplicate monster record union views

* reconstruct: consolidate army group type and signed count owners

* reconstruct: separate path directions from grid pointer storage

* reconstruct: simplify NetBIOS payload and event owners

* reconstruct: recover town state owners and natural setup assignments

* reconstruct: bound startup flags and audit retained frame storage

* reconstruct: simplify map-state conversions and audit hint retry exit

* reconstruct: simplify save/load conversions and player scratch

* reconstruct: recover named spell-widget selections

* reconstruct: recover map-input byte owners and assignments

* reconstruct: simplify sprite and hero initialization conversions

* reconstruct: simplify turn-update flag and scalar conversions

* reconstruct: recover new-turn text owner and audit unused locals

* reconstruct: simplify weekly casts and measure counter roles

* tools: select marked function before trailing inline helpers

* reconstruct: simplify NewMap assignments and audit structural alternatives

* reconstruct: clarify save transfer conversions and exit flag

* reconstruct: recover shared save wire types and payload extents

* reconstruct: simplify diff and codec casts with bounded evidence

* reconstruct: audit remaining GAME joins and simplify six casts

* reconstruct: audit town handler storage and simplify eight casts

+++ include/BASE/Misc.h
@@ -28,7 +28,7 @@ H2_ENUM_END(LogConstant)
 // Existing signed integer deltas; no widening or alternative distance metric.
 #define MANHATTAN_LENGTH(dx, dy) (abs((dx)) + abs((dy)))
 #define INTEGER_VECTOR_LENGTH(dx, dy) \
-    (static_cast<i32>(sqrt(static_cast<double>((dx) * (dx) + (dy) * (dy)))))
+    (static_cast<i32>(sqrt((dx) * (dx) + (dy) * (dy))))
 
 struct indexArray {
     u16 key;
